refactor: visualTransformation 컴포저블 제거 및 객체로 개선
