--- app/src/main/java/woowacourse/payments/ui/components/CreditCardVisualTransformation.kt
@@ -1,40 +1,0 @@
-package woowacourse.payments.ui.components
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.input.OffsetMapping
-import androidx.compose.ui.text.input.TransformedText
-import androidx.compose.ui.text.input.VisualTransformation
-
-@Composable
-fun CreditCardVisualTransformation(): VisualTransformation {
-    return VisualTransformation { text ->
-        val trimmed = if (text.text.length >= 16) text.text.substring(0..15) else text.text
-        val out = buildString {
-            for (i in trimmed.indices) {
-                append(trimmed[i])
-                if (i % 4 == 3 && i != 15) append("-")
-            }
-        }
-
-        val creditCardOffsetTranslator = object : OffsetMapping {
-            override fun originalToTransformed(offset: Int): Int {
-                if (offset <= 3) return offset
-                if (offset <= 7) return offset + 1
-                if (offset <= 11) return offset + 2
-                if (offset <= 16) return offset + 3
-                return out.length
-            }
-
-            override fun transformedToOriginal(offset: Int): Int {
-                if (offset <= 4) return offset
-                if (offset <= 9) return offset - 1
-                if (offset <= 14) return offset - 2
-                if (offset <= 19) return offset - 3
-                return trimmed.length
-            }
-        }
-
-        TransformedText(AnnotatedString(out), creditCardOffsetTranslator)
-    }
-}
